pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "locust-sa-test"
version = "0.1.0"
description = "Syncs Git repos across different providers and networks"
authors = ["CTS <https://cts.co>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
locust = "^2.10.0"
google-cloud-secret-manager = "^2.11.1"
google-auth = "^2.9.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
bandit = "^1.7.4"
pep8 = "^1.7.1"
pre-commit = "^2.19.0"

[tool.black]
line-length = 80

[tool.bandit]
skips = ["B101"]